XX-Workshop1/requirements.txt
google-cloud-bigquery-storage
dlt
dlt[duckdb]
dlt[bigquery]
dlt[pyiceberg]